third_party/WebKit/LayoutTests/inspector/elements/shadow/elements-panel-shadow-selection-on-refresh-expected.txt

Tests that elements panel preserves selected shadow DOM node on page refresh.


Running: setup

Running: testOpenShadowRoot
Page reloaded.
Selected node: '#shadow-root'

Running: testUserAgentShadowRoot
Page reloaded.
Selected node: '#shadow-root (user-agent)'

Running: testOpenShadowRootChild
Page reloaded.
Selected node: 'input'

Running: testUserAgentShadowRootChild
Page reloaded.
Selected node: 'div'
